quick fix + indowebnovel
